requirements.txt
Brotli==1.0.9
click==8.0.4
cycler==0.11.0
dash==2.2.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.3
Flask-Compress==1.10.1
fonttools==4.29.1
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.1
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.0
Pillow==9.0.1
pkg_resources==0.0.0
plotly==5.6.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
tenacity==8.0.1
Wave==0.0.2
Werkzeug==2.0.3
